+++ src/main/java/it/gov/pagopa/pu/classification/service/XMLUnmarshallerService.java
@@ -1,6 +1,6 @@
 package it.gov.pagopa.pu.classification.service;
 
-import it.gov.pagopa.pu.classification.exception.InvalidValueException;
+import it.gov.pagopa.pu.classification.exception.custom.InvalidValueException;
 import jakarta.xml.bind.JAXBContext;
 import jakarta.xml.bind.JAXBElement;
 import jakarta.xml.bind.JAXBException;
@@ -12,7 +12,6 @@ import org.springframework.stereotype.Service;
 import javax.xml.transform.stream.StreamSource;
 import javax.xml.validation.Schema;
 import java.io.ByteArrayInputStream;
-import java.io.FileInputStream;
 import java.io.IOException;
 import java.io.InputStream;
